% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic_model.R
\name{aic_model}
\alias{aic_model}
\title{aic_model
to be documented}
\usage{
aic_model(model)
}
\arguments{
\item{model}{to be documented}
}
\value{
to be documented
}
\description{
aic_model
to be documented
}
\keyword{internal}
