% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApproxiW.R
\name{ApproxiW}
\alias{ApproxiW}
\title{ApproxiW
to be documented}
\usage{
ApproxiW(WW, la, order)
}
\arguments{
\item{WW}{to be documented}

\item{la}{to be documented}

\item{order}{to be documented}
}
\value{
to be documented
}
\description{
ApproxiW
to be documented
}
\keyword{internal}
