% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_control.R
\name{assign_control}
\alias{assign_control}
\title{assign_control
to be documented}
\usage{
assign_control(control=list(),n, env = parent.frame())
}
\arguments{
\item{control}{to be documented}

\item{n}{to be documented}

\item{env}{env = parent.frame()}
}
\value{
to be documented
}
\description{
assign_control
to be documented
}
\keyword{internal}
