% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth_searchW.R
\name{bandwidth_searchW}
\alias{bandwidth_searchW}
\title{bandwidth_searchW
to be documented}
\usage{
bandwidth_searchW(formula,data,coord,fixed_vars,Model,control,kernel,kernel_w,e_search)
}
\arguments{
\item{formula}{to be documented}

\item{data}{to be documented}

\item{coord}{to be documented}

\item{fixed_vars}{to be documented}

\item{Model}{to be documented}

\item{control}{to be documented}

\item{kernel}{to be documented}

\item{kernel_w}{to be documented}

\item{e_search}{to be documented}
}
\value{
to be documented
}
\description{
bandwidth_searchW
to be documented
}
\keyword{internal}
