% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_gwrsar.R
\name{bias_gwrsar}
\alias{bias_gwrsar}
\title{bias_gwrsar
to be documented}
\usage{
bias_gwrsar(model, data, Model, DGP)
}
\arguments{
\item{model}{to be documented}

\item{data}{to be documented}

\item{Model}{to be documented}

\item{DGP}{to be documented}
}
\value{
to be documented
}
\description{
bias_gwrsar
to be documented
}
\keyword{internal}
