% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_C.R
\name{bin_C}
\alias{bin_C}
\title{bin_C
to be documented}
\usage{
bin_C(dd, hh, Minv)
}
\arguments{
\item{dd}{to be documented}

\item{hh}{to be documented}

\item{Minv}{to be documented}
}
\value{
to be documented
}
\description{
bin_C
to be documented
}
\keyword{internal}
