% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisq.R
\name{bisq}
\alias{bisq}
\title{bisquare kernel}
\usage{
bisq(d, h)
}
\arguments{
\item{d}{a vector of distance}

\item{h}{a distance bandwidth}
}
\value{
a vector of weight
}
\description{
bisquare kernel
}
\examples{
\donttest{
w=bisq(-30:30, 20)
plot(-30:30,w,type='l')
abline(v=-20)
abline(v=20)
}
}
