% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisq_C.R
\name{bisq_C}
\alias{bisq_C}
\title{bisquare kernel, RcppEigen version}
\usage{
bisq_C(d, h, Minv)
}
\arguments{
\item{d}{a vector of distance}

\item{h}{a distance bandwidth}

\item{Minv}{Minimal number of non null weight (neigbor)}
}
\value{
a vector of weight
}
\description{
bisquare kernel, RcppEigen version
}
\examples{
\donttest{
w=bisq_C(1:100,20,0)
plot(1:100,w,type='l')
}
}
