% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bisq_knn_C.R
\name{bisq_knn_C}
\alias{bisq_knn_C}
\title{Adaptive bisquare kernel}
\usage{
bisq_knn_C(d,h)
}
\arguments{
\item{d}{a vector of distance}

\item{h}{bandwidth size expressed in number of neighbors}
}
\value{
a vector of weight
}
\description{
Adaptive bisquare kernel
}
\examples{
\donttest{
w=bisq_knn_C(-30:30,20)
plot(-30:30,w,type='l')
abline(v=-10)
abline(v=10)
}
}
