% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP_pred_MGWRSAR.R
\name{BP_pred_MGWRSAR}
\alias{BP_pred_MGWRSAR}
\title{BP_pred_MGWRSAR
to be documented}
\usage{
BP_pred_MGWRSAR(YS,X,W,e,beta_hat,lambda_hat,S,O,coord,type='BPN',k=16,Wk=NULL)
}
\arguments{
\item{YS}{to be documented}

\item{X}{to be documented}

\item{W}{to be documented}

\item{e}{to be documented}

\item{beta_hat}{to be documented}

\item{lambda_hat}{to be documented}

\item{S}{to be documented}

\item{O}{to be documented}

\item{coord}{to be documented}

\item{type}{to be documented}

\item{k}{to be documented}

\item{Wk}{to be documented}
}
\value{
to be documented
}
\description{
BP_pred_MGWRSAR
to be documented
}
\keyword{internal}
