% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP_pred_MGWRSARC.R
\name{BP_pred_MGWRSARC}
\alias{BP_pred_MGWRSARC}
\title{BP_pred_MGWRSARC
to be documented}
\usage{
BP_pred_MGWRSARC(model,W,S,O,xnewG,varc,coords,type='BPN',k=16)
}
\arguments{
\item{model}{to be documented}

\item{W}{to be documented}

\item{S}{to be documented}

\item{O}{to be documented}

\item{xnewG}{to be documented}

\item{varc}{to be documented}

\item{coords}{to be documented}

\item{type}{to be documented}

\item{k}{to be documented}
}
\value{
to be documented
}
\description{
BP_pred_MGWRSARC
to be documented
}
\keyword{internal}
