% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BP_pred_SAR.R
\name{BP_pred_SAR}
\alias{BP_pred_SAR}
\title{BP_pred_SAR
to be documented}
\usage{
BP_pred_SAR(YS,X,W,e,beta_hat,lambda_hat,S,O,type='BPN',
model='SAR',W_extra=NULL,k_extra=30,kernel_extra='sheppard',coords=NULL)
}
\arguments{
\item{YS}{to be documented}

\item{X}{to be documented}

\item{W}{to be documented}

\item{e}{to be documented}

\item{beta_hat}{to be documented}

\item{lambda_hat}{to be documented}

\item{S}{to be documented}

\item{O}{to be documented}

\item{type}{to be documented}

\item{model}{to be documented}

\item{W_extra}{to be documented}

\item{k_extra}{to be documented}

\item{kernel_extra}{to be documented}

\item{coords}{to be documented}
}
\value{
to be documented
}
\description{
BP_pred_SAR
to be documented
}
\keyword{internal}
