% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef,mgwrsar-method}
\alias{coef,mgwrsar-method}
\title{coef for mgwrsar model}
\usage{
\S4method{coef}{mgwrsar}(object, ...)
}
\arguments{
\item{object}{A model of class \code{\link{mgwrsar-class}}.}

\item{...}{coef parameters forwarded.}
}
\value{
A named list with a matrix of varying coefficients and a vector or non varying coefficients.
}
\description{
coef for mgwrsar model
}
