% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_dense_C.R
\name{D_dense_C}
\alias{D_dense_C}
\title{D_dense_C
to be documented}
\usage{
D_dense_C(xxi, yyi, xx, yy)
}
\arguments{
\item{xxi}{to be documented}

\item{yyi}{to be documented}

\item{xx}{to be documented}

\item{yy}{to be documented}
}
\value{
to be documented
}
\description{
D_dense_C
to be documented
}
\keyword{internal}
