% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dx_dense_C.R
\name{Dx_dense_C}
\alias{Dx_dense_C}
\title{Dx_dense_C
to be documented}
\usage{
Dx_dense_C(xxi, xx, TIME)
}
\arguments{
\item{xxi}{to be documented}

\item{xx}{to be documented}

\item{TIME}{to be documented}
}
\value{
to be documented
}
\description{
Dx_dense_C
to be documented
}
\keyword{internal}
