% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDK.R
\name{EDK}
\alias{EDK}
\title{EDK
to be documented}
\usage{
EDK(d, h, k = 2)
}
\arguments{
\item{d}{to be documented}

\item{h}{to be documented}

\item{k}{to be documented}
}
\value{
to be documented
}
\description{
EDK
to be documented
}
\keyword{internal}
