% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastlmLLT_C.R
\name{fastlmLLT_C}
\alias{fastlmLLT_C}
\title{fastlmLLT_C
to be documented}
\usage{
fastlmLLT_C(XX, YY, SE_)
}
\arguments{
\item{XX}{to be documented}

\item{YY}{to be documented}

\item{SE_}{to be documented}
}
\value{
to be documented
}
\description{
fastlmLLT_C
to be documented
}
\keyword{internal}
