% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb.R
\name{fb}
\alias{fb}
\title{fb
to be documented}
\usage{
fb(formula,data,coord,fixed_vars,Model,control,kernel,e_search)
}
\arguments{
\item{formula}{to be documented}

\item{data}{to be documented}

\item{coord}{to be documented}

\item{fixed_vars}{to be documented}

\item{Model}{to be documented}

\item{control}{to be documented}

\item{kernel}{to be documented}

\item{e_search}{to be documented}
}
\value{
to be documented
}
\description{
fb
to be documented
}
\keyword{internal}
