% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_DGPTAB.R
\name{fill_DGPTAB}
\alias{fill_DGPTAB}
\title{fill_DGPTAB
to be documented}
\usage{
fill_DGPTAB(formula,data,coord,fixed_vars,Model,control,opt,search_W,config_model)
}
\arguments{
\item{formula}{to be documented}

\item{data}{to be documented}

\item{coord}{to be documented}

\item{fixed_vars}{to be documented}

\item{Model}{to be documented}

\item{control}{to be documented}

\item{opt}{to be documented}

\item{search_W}{to be documented}

\item{config_model}{to be documented}
}
\value{
to be documented
}
\description{
fill_DGPTAB
to be documented
}
\keyword{internal}
