% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted,mgwrsar-method}
\alias{fitted,mgwrsar-method}
\title{fitted for mgwrsar model}
\usage{
\S4method{fitted}{mgwrsar}(object, ...)
}
\arguments{
\item{object}{A model of class \code{\link{mgwrsar-class}}.}

\item{...}{fitted parameters forwarded.}
}
\value{
A vector of fitted values.
}
\description{
fitted for mgwrsar model
}
