% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_adapt.R
\name{gauss_adapt}
\alias{gauss_adapt}
\title{Adaptive gaussian kernel}
\usage{
gauss_adapt(d, h)
}
\arguments{
\item{d}{a vector of distance}

\item{h}{bandwidth size expressed in number of neighbors}
}
\value{
a vector of weight
}
\description{
Adaptive gaussian kernel
}
\examples{
\donttest{
w=gauss_adapt(-30:30,20)
plot(-30:30,w,type='l')
abline(v=-10)
abline(v=10)
}
}
