% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_C.R
\name{gauss_C}
\alias{gauss_C}
\title{gauss_C
to be documented}
\usage{
gauss_C(dd, hh, Minv)
}
\arguments{
\item{dd}{to be documented}

\item{hh}{to be documented}

\item{Minv}{to be documented}
}
\value{
to be documented
}
\description{
gauss_C
to be documented
}
\keyword{internal}
