% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_knn_C.R
\name{gauss_knn_C}
\alias{gauss_knn_C}
\title{gauss_knn_C
to be documented}
\usage{
gauss_knn_C(dd, hh)
}
\arguments{
\item{dd}{to be documented}

\item{hh}{to be documented}
}
\value{
to be documented
}
\description{
gauss_knn_C
to be documented
}
\keyword{internal}
