% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPKj.R
\name{GPKj}
\alias{GPKj}
\title{GPKj
to be documented}
\usage{
GPKj(j, Y,X,S, H, kernels, type = "GD", minv = 0,
maxknn = 500, NmaxDist = 6000, isgcv, TIME, decay)
}
\arguments{
\item{j}{to be documented}

\item{Y}{to be documented}

\item{X}{to be documented}

\item{S}{to be documented}

\item{H}{to be documented}

\item{kernels}{to be documented}

\item{type}{to be documented}

\item{minv}{to be documented}

\item{maxknn}{to be documented}

\item{NmaxDist}{to be documented}

\item{isgcv}{to be documented}

\item{TIME}{to be documented}

\item{decay}{to be documented}
}
\value{
to be documented
}
\description{
GPKj
to be documented
}
\keyword{internal}
