% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incr_search_band.R
\name{incr_search_band}
\alias{incr_search_band}
\title{incr_search_band
to be documented}
\usage{
incr_search_band(f,lower,discrete=FALSE,n,...)
}
\arguments{
\item{f}{to be documented}

\item{lower}{to be documented}

\item{discrete}{to be documented}

\item{n}{to be documented}

\item{...}{to be documented}
}
\value{
to be documented
}
\description{
incr_search_band
to be documented
}
\keyword{internal}
