% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indice_aggreg.R
\name{indice_aggreg}
\alias{indice_aggreg}
\title{indice_aggreg
to be documented}
\usage{
indice_aggreg(D)
}
\arguments{
\item{D}{to be documented}
}
\value{
to be documented
}
\description{
indice_aggreg
to be documented
}
\keyword{internal}
