% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INST_C.R
\name{INST_C}
\alias{INST_C}
\title{INST_C
to be documented}
\usage{
INST_C(XX, WW, withlambda, llambda)
}
\arguments{
\item{XX}{to be documented}

\item{WW}{to be documented}

\item{withlambda}{to be documented}

\item{llambda}{to be documented}
}
\value{
to be documented
}
\description{
INST_C
to be documented
}
\keyword{internal}
