% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_C.R
\name{kernel_C}
\alias{kernel_C}
\title{kernel_C
to be documented}
\usage{
kernel_C(XX, J, hh, Mykernel, Minv, TIME, Decay, DDiagNull, normWW)
}
\arguments{
\item{XX}{to be documented}

\item{J}{to be documented}

\item{hh}{to be documented}

\item{Mykernel}{to be documented}

\item{Minv}{to be documented}

\item{TIME}{to be documented}

\item{Decay}{to be documented}

\item{DDiagNull}{to be documented}

\item{normWW}{to be documented}
}
\value{
to be documented
}
\description{
kernel_C
to be documented
}
\keyword{internal}
