% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{KNN}
\alias{KNN}
\title{A function that returns a row normalized weight matrix based on k first neighbors,
to be documented}
\usage{
KNN(coord,h,diagnull=TRUE,kernel='knn',query=NULL)
}
\arguments{
\item{coord}{matrix of coordinates}

\item{h}{a bandwidth}

\item{diagnull}{0 on diagonal, default TRUE}

\item{kernel}{kernel type ('bisq','bisq_knn','gauss','gauss_adapt','knn')}

\item{query}{an index of neighbors to consider, if NULL all observation are used.}
}
\value{
a row nomralized weight dgCmatrix
}
\description{
A function that returns a row normalized weight matrix based on k first neighbors,
to be documented
}
\examples{
\donttest{
data(mydata)
coord=as.matrix(mydata[,c("x_lat","y_lon")])
W=KNN(coord,8)
which(W[1,]>0)
W[1,W[1,]>0]
}
}
