% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNNX.R
\name{KNNX}
\alias{KNNX}
\title{KNNX
to be documented}
\usage{
KNNX(X,Z=NULL,k_neighbors=NULL,h_bandwidth=NULL,diagnull=TRUE,
kernel='gauss_adapt',query=NULL,Zquery=NULL,ZC1=NULL,ZC2=NULL,
conditionZ=NULL,hd=0,norm=TRUE)
}
\arguments{
\item{X}{to be documented}

\item{Z}{to be documented}

\item{k_neighbors}{to be documented}

\item{h_bandwidth}{to be documented}

\item{diagnull}{to be documented}

\item{kernel}{to be documented}

\item{query}{to be documented}

\item{Zquery}{to be documented}

\item{ZC1}{to be documented}

\item{ZC2}{to be documented}

\item{conditionZ}{to be documented}

\item{hd}{to be documented}

\item{norm}{to be documented}
}
\value{
to be documented
}
\description{
KNNX
to be documented
}
\keyword{internal}
