% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_gwrsar.R
\name{me_gwrsar}
\alias{me_gwrsar}
\title{me_gwrsar
to be documented}
\usage{
me_gwrsar(model, data, W_hat = NULL, Wtrue = NULL, DGP, Model, verbose = TRUE)
}
\arguments{
\item{model}{to be documented}

\item{data}{to be documented}

\item{W_hat}{to be documented}

\item{Wtrue}{to be documented}

\item{DGP}{to be documented}

\item{Model}{to be documented}

\item{verbose}{to be documented}
}
\value{
to be documented
}
\description{
me_gwrsar
to be documented
}
\keyword{internal}
