% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgwrsar_bootstrap_test.R
\name{mgwrsar_bootstrap_test}
\alias{mgwrsar_bootstrap_test}
\title{A bootstrap test for Betas for mgwrsar class model.}
\usage{
mgwrsar_bootstrap_test(x0,x1,B=100,ncore=1,type='standard',eps='H1',
df='H1',focal='median',D=NULL,verbose=FALSE)
}
\arguments{
\item{x0}{The H0 mgwrsar model}

\item{x1}{The H1 mgwrsar model}

\item{B}{number of bootstrap repetitions, default 100}

\item{ncore}{number of cores}

\item{type}{type of bootstap : 'wild','Rademacher','spatial' or 'standard' (default)}

\item{eps}{Hypothesis under wich residuals are simulated,  'H0' or 'H1' (default)}

\item{df}{Hypothesis under wich degree of freedom is estimated.}

\item{focal}{see sample_stat help}

\item{D}{A matrix of distance}

\item{verbose}{default FAlSE}
}
\value{
The value of the statistics test and a p ratio.
}
\description{
A bootstrap test for Betas for mgwrsar class model.
}
\seealso{
mgwrsar_bootstrap_test_all
}
