% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgwrsar_bootstrap_test_all.R
\name{mgwrsar_bootstrap_test_all}
\alias{mgwrsar_bootstrap_test_all}
\title{A bootstrap test for testing nullity of all Betas for mgwrsar class model,}
\usage{
mgwrsar_bootstrap_test_all(model,B=100,doMC=FALSE,ncore=1,
type='standard')
}
\arguments{
\item{model}{A mgwrsar model}

\item{B}{number of bootstrap replications, default 100}

\item{doMC}{If TRUE, doMC parallelization}

\item{ncore}{number of cores.}

\item{type}{type of boostrap ('spatial','wild','random')}
}
\value{
a matrix with statistical test values and p ratios
}
\description{
A bootstrap test for testing nullity of all Betas for mgwrsar class model,
}
\seealso{
mgwrsar_bootstrap_test
}
