% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{mod}
\alias{mod}
\title{mod
to be documented}
\usage{
mod(YY, XX, WW, XZZ, YZZ, Wi, LocalInst, ismethodB2SLS, ismethodMGWRSAR_1_kc_0, SE_)
}
\arguments{
\item{YY}{to be documented}

\item{XX}{to be documented}

\item{WW}{to be documented}

\item{XZZ}{to be documented}

\item{YZZ}{to be documented}

\item{Wi}{to be documented}

\item{LocalInst}{to be documented}

\item{ismethodB2SLS}{to be documented}

\item{ismethodMGWRSAR_1_kc_0}{to be documented}

\item{SE_}{to be documented}
}
\value{
to be documented
}
\description{
mod
to be documented
}
\keyword{internal}
