% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modc.R
\name{modc}
\alias{modc}
\title{modc is a set of models to correct approximation of hat matrix trace}
\description{
modc is a set of models to correct approximation of hat matrix trace
}
\references{
\doi{10.1016/j.regsciurbeco.2017.04.001}
}
\author{
Ghislain Geniaux and Davide Martinetti \email{ghislain.geniaux@inra.fr}
}
