% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_conf.R
\name{model_conf}
\alias{model_conf}
\title{model_conf
to be documented}
\usage{
model_conf(formula,data,coord,fixed_vars,Model,control,config_model,search_W)
}
\arguments{
\item{formula}{to be documented}

\item{data}{to be documented}

\item{coord}{to be documented}

\item{fixed_vars}{to be documented}

\item{Model}{to be documented}

\item{control}{to be documented}

\item{config_model}{to be documented}

\item{search_W}{to be documented}
}
\value{
to be documented
}
\description{
model_conf
to be documented
}
\keyword{internal}
