% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.multiscale_gwr.R
\name{multiscale_gwr.cv}
\alias{multiscale_gwr.cv}
\title{multiscale_gwr.cv
to be documented (experimental)}
\usage{
multiscale_gwr.cv(dataName, argDataName="data", target='Y', K=5, regFun, par_model,
par_model2=NULL,regFun2=NULL, predFun, args_predNames, extra_args_pred=NULL,
namesXtraArgs2Split=NULL,myseed=1)
}
\arguments{
\item{dataName}{character, name of the data}

\item{argDataName}{character, generic name to use as data name.}

\item{target}{character, name of variable to explain}

\item{K}{integer, number of folds for cross validation}

\item{regFun}{character, name of the estimation function}

\item{par_model}{named list with the arguments for the estimation function}

\item{par_model2}{to be documented}

\item{regFun2}{to be documented}

\item{predFun}{character, name of the prediction function}

\item{args_predNames}{named list with the arguments for the prediction function}

\item{extra_args_pred}{named list with extra arguments for non generic prediction function}

\item{namesXtraArgs2Split}{character, names of the objects in extra_args_pred that need to be split for cross validation.}

\item{myseed}{seed for random number.}
}
\description{
multiscale_gwr.cv
to be documented (experimental)
}
