% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mydata}
\alias{mydata}
\title{mydata is a simulated data set of a mgwrsar model}
\format{
A data frames with 1000 rows 22 variables and a matrix of coordinates with two columns
}
\description{
mydata is a simulated data set of a mgwrsar model
}
\references{
\doi{10.1016/j.regsciurbeco.2017.04.001}
}
\author{
Ghislain Geniaux and Davide Martinetti \email{ghislain.geniaux@inrae.fr}
}
