% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mydatasf}
\alias{mydatasf}
\title{mydataf is a Simple Feature object with real estate data in south of France.}
\format{
A sf object with 1403 rows, 5 columns
}
\description{
mydataf is a Simple Feature object with real estate data in south of France.
}
\references{
\url{https://www.data.gouv.fr/fr/datasets/demandes-de-valeurs-foncieres/}
}
\author{
Ghislain Geniaux \email{ghislain.geniaux@inrea.fr}
}
