% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhWY_C.R
\name{PhWY_C}
\alias{PhWY_C}
\title{PhWY_C
to be documented}
\usage{
PhWY_C(YY, XX, WW, Wi)
}
\arguments{
\item{YY}{to be documented}

\item{XX}{to be documented}

\item{WW}{to be documented}

\item{Wi}{to be documented}
}
\value{
to be documented
}
\description{
PhWY_C
to be documented
}
\keyword{internal}
