% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proj_C.R
\name{Proj_C}
\alias{Proj_C}
\title{Proj_C
to be documented}
\usage{
Proj_C(HH, XX)
}
\arguments{
\item{HH}{to be documented}

\item{XX}{to be documented}
}
\value{
to be documented
}
\description{
Proj_C
to be documented
}
\keyword{internal}
