% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRcpp2_C.R
\name{QRcpp2_C}
\alias{QRcpp2_C}
\title{QRcpp2_C
to be documented}
\usage{
QRcpp2_C(AA, bb, cc)
}
\arguments{
\item{AA}{to be documented}

\item{bb}{to be documented}

\item{cc}{to be documented}
}
\value{
to be documented
}
\description{
QRcpp2_C
to be documented
}
\keyword{internal}
