% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRcpp_C.R
\name{QRcpp_C}
\alias{QRcpp_C}
\title{QRcpp_C
to be documented}
\usage{
QRcpp_C(AA, bb)
}
\arguments{
\item{AA}{to be documented}

\item{bb}{to be documented}
}
\value{
to be documented
}
\description{
QRcpp_C
to be documented
}
\keyword{internal}
