% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reord_D.R
\name{reord_D}
\alias{reord_D}
\title{reord_D}
\usage{
reord_D(M, idxG)
}
\arguments{
\item{M}{A numeric matrix to be reordered.}

\item{idxG}{An integer matrix of the same number of rows as `M`, where
each row contains column indices indicating the new ordering for that row.}
}
\value{
A reordered numeric matrix with the same number of rows as `M` and
columns equal to the maximum index in `idxG`.
}
\description{
Reorders the elements of a distance matrix row by row according to a corresponding
index matrix.
}
