% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals,mgwrsar-method}
\alias{residuals,mgwrsar-method}
\title{residuals for mgwrsar model}
\usage{
\S4method{residuals}{mgwrsar}(object, ...)
}
\arguments{
\item{object}{A model of class \code{\link{mgwrsar-class}}.}

\item{...}{residuals parameters forwarded.}
}
\value{
A vector of residuals.
}
\description{
residuals for mgwrsar model
}
