% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_spat.R
\name{sample_spat}
\alias{sample_spat}
\title{A sampling method for bootstrap test with spatial data using distance bisquare kernel.}
\usage{
sample_spat(D,h,focal='median')
}
\arguments{
\item{D}{A square matrix of distance.}

\item{h}{A distance bandwidth.}

\item{focal}{If focal is NULL, the ith residuals can not be selected for the ith observation, if focal=='median' it could be selected with a probability equals to the median probability of neighbors.}
}
\value{
a vector of bootstrapped index
}
\description{
A sampling method for bootstrap test with spatial data using distance bisquare kernel.
}
\keyword{internal}
