% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SARHS.R
\name{SARHS}
\alias{SARHS}
\title{SARHS
to be documented}
\usage{
SARHS(Xw,YYw,SE,W,Betav)
}
\arguments{
\item{Xw}{to be documented}

\item{YYw}{to be documented}

\item{SE}{to be documented}

\item{W}{to be documented}

\item{Betav}{to be documented}
}
\value{
to be documented
}
\description{
SARHS
to be documented
}
\keyword{internal}
