% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set0_conditionXD.R
\name{set0_conditionXD}
\alias{set0_conditionXD}
\title{set0_conditionXD
to be documented}
\usage{
set0_conditionXD(XX, ZZ)
}
\arguments{
\item{XX}{to be documented}

\item{ZZ}{to be documented}
}
\value{
to be documented
}
\description{
set0_conditionXD
to be documented
}
\keyword{internal}
