% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sidak_cor_MGWR.R
\name{Sidak_cor_MGWR}
\alias{Sidak_cor_MGWR}
\title{Sidak_cor_MGWR
to be documented}
\usage{
Sidak_cor_MGWR(alpha=0.05,modelH0,modelH1,D=NULL,type='rn')
}
\arguments{
\item{alpha}{to be documented}

\item{modelH0}{to be documented}

\item{modelH1}{to be documented}

\item{D}{to be documented}

\item{type}{to be documented}
}
\value{
to be documented
}
\description{
Sidak_cor_MGWR
to be documented
}
\keyword{internal}
