% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sl_C.R
\name{Sl_C}
\alias{Sl_C}
\title{Sl_C
to be documented}
\usage{
Sl_C(llambda, WW, iinv, aapprox)
}
\arguments{
\item{llambda}{to be documented}

\item{WW}{to be documented}

\item{iinv}{to be documented}

\item{aapprox}{to be documented}
}
\value{
to be documented
}
\description{
Sl_C
to be documented
}
\keyword{internal}
