% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSR_h.R
\name{SSR_h}
\alias{SSR_h}
\title{SSR_h
to be documented}
\usage{
SSR_h(formula, data,  coord, fixed_vars,kernels, H, Model, control=list(),Penalized)
}
\arguments{
\item{formula}{to be documented}

\item{data}{to be documented}

\item{coord}{to be documented}

\item{fixed_vars}{to be documented}

\item{kernels}{to be documented}

\item{H}{to be documented}

\item{Model}{to be documented}

\item{control}{to be documented}

\item{Penalized}{to be documented}
}
\value{
to be documented
}
\description{
SSR_h
to be documented
}
\keyword{internal}
