% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSR_h_bw.R
\name{SSR_h_bw}
\alias{SSR_h_bw}
\title{SSR_h_bw
to be documented}
\usage{
SSR_h_bw(formula, data,  coord, fixed_vars,kernels, H,
h_w,kernel_w='bisq',W1=NULL, Model, control=list(),Penalized)
}
\arguments{
\item{formula}{to be documented}

\item{data}{to be documented}

\item{coord}{to be documented}

\item{fixed_vars}{to be documented}

\item{kernels}{to be documented}

\item{H}{to be documented}

\item{h_w}{to be documented}

\item{kernel_w}{to be documented}

\item{W1}{to be documented}

\item{Model}{to be documented}

\item{control}{to be documented}

\item{Penalized}{to be documented}
}
\value{
to be documented
}
\description{
SSR_h_bw
to be documented
}
\keyword{internal}
