% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_search.R
\name{support_search}
\alias{support_search}
\title{support_search
to be documented}
\usage{
support_search(formula, data,  coord, fixed_vars, Model,control,kernel,e_search)
}
\arguments{
\item{formula}{to be documented}

\item{data}{to be documented}

\item{coord}{to be documented}

\item{fixed_vars}{to be documented}

\item{Model}{to be documented}

\item{control}{to be documented}

\item{kernel}{to be documented}

\item{e_search}{to be documented}
}
\value{
to be documented
}
\description{
support_search
to be documented
}
\keyword{internal}
