% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tf.R
\name{Tf}
\alias{Tf}
\title{Asymptotic test for mgwrsar models using Sidak correction, EXPERIMENTAL (alpha=0.05 percent)}
\usage{
Tf(model_A,model_B,pv=FALSE)
}
\arguments{
\item{model_A}{A mgwrsar model H0}

\item{model_B}{A mgwrsar model H1}

\item{pv}{If TRUE p-value are estmimated using Sidak correction, default FALSE}
}
\value{
to be documented
}
\description{
Asymptotic test for mgwrsar models using Sidak correction, EXPERIMENTAL (alpha=0.05 percent)
}
\keyword{internal}
