% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_search_bandwidth_AICc.R
\name{golden_search_bandwidth}
\alias{golden_search_bandwidth}
\title{golden_search_bandwidth
to be documented}
\usage{
golden_search_bandwidth(formula,H2=NULL,data, coords, fixed_vars,
kernels, Model, control,lower.bound, upper.bound,tolerance=0.000001)
}
\arguments{
\item{formula}{to be documented}

\item{H2}{to be documented}

\item{data}{to be documented}

\item{coords}{to be documented}

\item{fixed_vars}{to be documented}

\item{kernels}{to be documented}

\item{Model}{to be documented}

\item{control}{to be documented}

\item{lower.bound}{to be documented}

\item{upper.bound}{to be documented}

\item{tolerance}{to be documented}
}
\value{
a list(minimum=res,objective=objective,model=model).
}
\description{
golden_search_bandwidth
to be documented
}
