% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_prems.R
\name{int_prems}
\alias{int_prems}
\title{int_prems
to be documented}
\usage{
int_prems(X)
}
\arguments{
\item{X}{to be documented}
}
\value{
to be documented
}
\description{
int_prems
to be documented
}
\keyword{internal}
