% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiscale_gwr.R
\name{multiscale_gwr}
\alias{multiscale_gwr}
\title{multiscale_gwr
to be documented (experimental)}
\usage{
multiscale_gwr(formula, data,coord, fixed_vars=NULL,Model=c('GWR'),
kernels=c('bisq'),maxiter=10,control=list(),fd=c(0.6,0.9,0.99,1,1.01,1.1,1.4),
H=NULL,BETA=NULL,grad=FALSE,nstable=3,nfull=2,init='GWR')
}
\arguments{
\item{formula}{a formula.}

\item{data}{a dataframe or a spatial dataframe (sp package).}

\item{coord}{default NULL, a dataframe or a matrix with coordinates, not
required if data is a spatial dataframe.}

\item{fixed_vars}{a vector with the names of spatiallay constant coefficient for
mixed model. All other variables present in formula are supposed to be spatially
varying. If empty or NULL (default), all variables in formula are supposed to be
spatially varying.}

\item{Model}{character containing the type of model: Possible values are "OLS",
"SAR", "GWR" (default), "MGWR" , "MGWRSAR_0_0_kv","MGWRSAR_1_0_kv",
"MGWRSAR_0_kc_kv", "MGWRSAR_1_kc_kv", "MGWRSAR_1_kc_0". See Details for more
explanation.}

\item{kernels}{A vector containing the kernel types. Possible types:
rectangle ("rectangle"), bisquare ("bisq"), tricube ("tcub"), epanechnikov ("epane"), gaussian
("gauss")) .}

\item{maxiter}{maximum number of backfiting iteration}

\item{control}{list of extra control arguments for MGWRSAR wrapper - see Details below}

\item{fd}{to be documented (experimental)}

\item{H}{vector containing the bandwidth parameters for each covariate.}

\item{BETA}{to be documented (experimental)}

\item{grad}{to be documented (experimental)}

\item{nstable}{to be documented (experimental)}

\item{nfull}{to be documented (experimental)}

\item{init}{to be documented (experimental)}
}
\description{
multiscale_gwr
to be documented (experimental)
}
