% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mydata}
\alias{mydata}
\title{mydata is a simulated data set of a mgwrsar model}
\format{An object of class \code{data.frame} with 1000 rows and 22 columns.}
\usage{
mydata
}
\description{
mydata is a simulated data set of a mgwrsar model
}
\references{
\url{https://www.sciencedirect.com/science/article/pii/S0166046216302381}
}
\author{
Ghislain Geniaux and Davide Martinetti \email{ghislain.geniaux@inra.fr}
}
\keyword{datasets}
