% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normW.R
\name{normW}
\alias{normW}
\title{normW
row normalization of dgCMatrix}
\usage{
normW(x)
}
\arguments{
\item{x}{A dgCMatrix class matrix}
}
\value{
A row normalized dgCMatrix
}
\description{
normW
row normalization of dgCMatrix
}
