% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normW.R
\name{normW}
\alias{normW}
\title{row normalization of dgCMatrix}
\usage{
normW(W)
}
\arguments{
\item{W}{A dgCMatrix class matrix}
}
\value{
A row normalized dgCMatrix
}
\description{
row normalization of dgCMatrix
}
\keyword{internal}
