% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot,mgwrsar,missing-method}
\alias{plot,mgwrsar,missing-method}
\alias{plot.mgwrsar}
\title{Plot method for mgwrsar model}
\usage{
\S4method{plot}{mgwrsar,missing}(
  x,
  y,
  type = "coef",
  var = NULL,
  crs = NULL,
  mypalette = "RdYlGn",
  opacity = 0.5,
  fopacity = 0.5,
  nbins = 8,
  radius = 500,
  mytile = "Stadia.StamenTonerBackground",
  myzoom = 8,
  myresolution = 150,
  LayersControl = TRUE,
  myzoomControl = TRUE,
  mytile2 = NULL,
  ScaleBar = NULL,
  ScaleBarOptions = list(maxWidth = 200, metric = TRUE, imperial = FALSE, updateWhenIdle
    = TRUE),
  MyLegendTitle = NULL,
  lopacity = 0.5
)
}
\arguments{
\item{x}{A model of class \code{\link{mgwrsar-class}}.}

\item{y}{missing}

\item{type}{default 'coef', for plotting the value of the coefficients. Local t-Student could also be plot using 't_coef', residuals using 'residuals' and fitted using 'fitted'.}

\item{var}{Names of variable to plot.}

\item{crs}{A CRS projection.}

\item{mypalette}{A leaflet palette.}

\item{opacity}{Opacity of border color.}

\item{fopacity}{Opacity of fill color.}

\item{nbins}{nbins.}

\item{radius}{radius of circle for plot of points.}

\item{mytile}{tile 1.}

\item{myzoom}{level of zoom for tile 1.}

\item{myresolution}{resolution for tile 1.}

\item{LayersControl}{layers contols.}

\item{myzoomControl}{zoem control.}

\item{mytile2}{tile 2.}

\item{ScaleBar}{ScaleBar.}

\item{ScaleBarOptions}{options for ScaleBar.}

\item{MyLegendTitle}{Legend title.}

\item{lopacity}{opacity for legend.}
}
\value{
A Interactive Web Maps with local parameters plot and Open Street Map layer.
}
\description{
Plot method for mgwrsar model
}
