% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary,mgwrsar-method}
\alias{summary,mgwrsar-method}
\title{summary for mgwrsar model}
\usage{
\S4method{summary}{mgwrsar}(object, ...)
}
\arguments{
\item{object}{A model of class \code{\link{mgwrsar-class}}.}

\item{...}{summary parameters forwarded.}
}
\value{
A summary object.
}
\description{
summary for mgwrsar model
}
