% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_Matrix.R
\name{summary_Matrix}
\alias{summary_Matrix}
\title{summary_Matrix
to be documented}
\usage{
summary_Matrix(object, ...)
}
\arguments{
\item{object}{to be documented}

\item{...}{to be documented}
}
\value{
to be documented
}
\description{
summary_Matrix
to be documented
}
