% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_mgwrsar.R
\name{summary_mgwrsar}
\alias{summary_mgwrsar}
\title{Print a summary of mgwrsar models}
\usage{
summary_mgwrsar(model)
}
\arguments{
\item{model}{a model of class mgwrsar}
}
\value{
a summary of mgwrsar models
}
\description{
Print a summary of mgwrsar models
}
\examples{
\dontrun{
library(mgwrsar)
data(data_mgwrsar)
coord=as.matrix(mydata[,c("x_lat","y_lon")])
W=KNN(coord,8)
model_GWR<-MGWRSAR(formula = 'Y_gwr~X1+X2+X3', data = mydata,coord=coord,
 fixed_vars=NULL,kernels=c('gauss'),H=0.13, Model = 'GWR',
 control=list(SE=TRUE))
summary_mgwrsar(model_GWR)
}
}
\seealso{
MGWRSAR, bandwidths_mgwrsar, plot_mgwrsar, predict_mgwrsar, kernelW_C
}
